xMsg: only destroy sockets if context is shadow
Don't call destroy when using the global context singleton.

+++ src/org/jlab/coda/xmsg/net/xMsgConnectionFactory.java
@@ -118,7 +118,9 @@ public class xMsgConnectionFactory {
     }
 
     public void destroy() {
-        context.destroy();
+        if (!context.isMain()) {
+            context.destroy();
+        }
     }
